# Add community version controlnet performance testing