--- a/onediff_comfy_nodes/README.md
+++ b/onediff_comfy_nodes/README.md
@@ -4,14 +4,23 @@
 <img src="../imgs/onediff_logo.png" height="100">
 </p>
 
-Performance of Community Edition
+---
+
+Performance of Community Edition 
 
 Updated on January 23, 2024. Device: RTX 3090
 
-<!-- | SDXL1.0-base (1024x1024)                                       | torch(Baseline) | onediff(Optimized) | Percentage improvement |
-| -------------------------------------------------------------- | --------------- | ------------------ | ---------------------- |
-| [Stable Diffusion workflow(UNet)](workflows/model-speedup.png) | 4.08it/s        | 6.70it/s           | 64.2 %                 |
-| [LoRA workflow](workflows/model-speedup-lora.png)              | 4.05it/s        | 6.69it/s           | 65.1 %                 | -->
+**SDXL End2End Time , Image Size 1024x1024, Batch Size 1 , steps 20**
+
+
+<img src="../imgs/comfy_community_version_speed.png">
+
+<details>
+<summary> Figure Notes </summary>
+
+- TODO: add workflow link
+
+</details>
 
 ## Documentation