## Performance Comparison
> NVIDIA GeForce RTX 3090 

- model: sd_xl_base_1.0.safetensors
- batch size: 1
- image size: 1024x1024
- steps: 20


Timings for 20 steps at 1024x1024
| e2e                    | Basic  | Lora   | ControlNet |
| ---------------------- | ------ | ------ | ---------- |
| PyTorch                | 5.57 s | 5.62 s | 7.95 s     |
| OneDiff                | 3.16 s | 3.16 s | 5.45 s     |
| Percentage improvement | 43%    | 44%    | 31%        |



Image throughput for 20 steps at 1024x1024
| it/s                   | Basic    | Lora     | ControlNet |
| ---------------------- | -------- | -------- | ---------- |
| PyTorch                | 4.04it/s | 3.99it/s | 2.87it/s   |
| OneDiff                | 7.61it/s | 7.51it/s | 4.40it/s   |
| Percentage improvement | 88%      | 88%      | 53%        |


- onediff: d387e7e9 (origin/main, origin/HEAD, main) 
- oneflow: version: 0.9.1.dev20240120+cu121 git_commit: 287b806